feat: Fix and report battery status * Update espp to have fixed switch pro input report battery / connection status byte * Update switch pro / gamepad device to support setting the battery level if * Update to subscribe to the battery level characteristic of the battery service and store the value when it changes - sending it in the usb gamepad report * Update to read the remote device serial number if it exists * Update to show the remote device serial number on the screen if it exists and the device has a screen
Improves usability of the device by showing the battery level of the controller on the switch (as best it can) as well as showing on the screen which controller it is connected to for easier use / debugging.

Build and run `main` on t-dongle-s3 and testing with Switch to ensure battery level is reported properly and the serial number shows up on the dongle as expected.